# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "This is a API to manage a Mechanic Shop"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Login endpoint, (generates token)"
      description: "Validates customer credentials, generates and return a customer specific token."
      parameters:
        - in: "body"
          name: "body"
          description: "Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: "Get a single customer by ID"
      description: "Endpoint that returns a specific customer from the database by their ID."
      parameters:
        - in: "path"
          name: "customer_id"
          description: "ID of the customer to retrieve"
          required: true
          type: "integer"
      responses:
        200:
          description: "customer found"
          schema:
            $ref: "#/definitions/AllCustomers"
        404:
          description: "Customer not found"
  /customers:
    post:
      tags:
        - Customers
      summary: "Endpoint to create/register a new user."
      description: "Endpoint to create a new customer in our Mechanic Shop system, emails must be unique."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new Customer."
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Successful Customer Creation"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: John Doe
              phone: "555-0101"
              email: john.doe@email.com
              password: password123

    get:
      tags:
        - Customers
      summary: "Returns all Customers"
      description: "Returns a list of all customers in our Mechanic Shop system."
      responses:
        200:
          description: "Successful retrieval of all customers"
          schema:
            $ref: "#/definitions/AllCustomers"

    put:
      tags:
        - Customers
      summary: "Endpoint to update a customer"
      description: "Endpoint to update a customer, and this is a token authenticated route, and the user must be logged in to update their account."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to update a Customer."
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Successful Customer Update"
          schema:
            $ref: "#/definitions/UpdateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: John Doe
              phone: "555-0101"
              email: john.doe@email.com
              password: password123
        404:
          description: "Customer not found"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              message: "Customer not found"
              status: "error"
              token: null

    delete:
      tags:
        - Customers
      summary: "Endpoint to delete a customer"
      description: "Endpoint to delete a customer, and this is a token authenticated route, and the user must be logged in to delete their account."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Successful Customer Deletion"
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
          examples:
            application/json:
              message: "Customer <customer_id> deleted successfully"
              status: "success"

# INVENTORY ENDPOINTS
...
definitions:
...
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"
    required:
      - email
      - password
  LoginResponse:
    type: "object"
    properties:
      message:
        type: "string"
      status:
        type: "string"
      token:
        type: "string"
  CreateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
    required:
      - name
      - email
      - phone
      - password
  CreateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  AllCustomers:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
        password:
          type: "string"
  UpdateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
    required:
      - name
      - email
      - phone
      - password
  UpdateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  DeleteCustomerResponse:
    type: "object"
    properties:
      message:
        type: "string"

# INVENTORY SCHEMAS
...